+++ src/main/java/de/markusbordihn/easynpc/data/WorldPresetData.java
@@ -1,0 +1,126 @@
+/**
+ * Copyright 2023 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package de.markusbordihn.easynpc.data;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.storage.LevelResource;
+
+import net.minecraftforge.server.ServerLifecycleHooks;
+
+import de.markusbordihn.easynpc.Constants;
+import de.markusbordihn.easynpc.data.skin.SkinModel;
+
+public class WorldPresetData {
+
+  protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+  protected static final String DATA_FOLDER_NAME = "preset";
+  public static final File MOD_FOLDER =
+      new File(ServerLifecycleHooks.getCurrentServer().getWorldPath(LevelResource.ROOT).toFile(),
+          Constants.MOD_ID);
+
+  private static ConcurrentHashMap<ResourceLocation, Path> presetResourceLocationMap =
+      new ConcurrentHashMap<>();
+
+  protected WorldPresetData() {}
+
+  public static void prepareFolder() {
+    log.info("{} world preset data ...", Constants.LOG_REGISTER_PREFIX);
+  }
+
+  public static Path getPresetDataFolder() {
+    Path path = MOD_FOLDER.toPath().resolve(DATA_FOLDER_NAME);
+    if (!path.toFile().exists()) {
+      path.toFile().mkdirs();
+    }
+    return path;
+  }
+
+  public static Path getPresetDataFolder(SkinModel skinModel) {
+    Path presetDataFolder = getPresetDataFolder();
+    String skinModelName = skinModel.name();
+    if (skinModelName != null) {
+      Path path = presetDataFolder.resolve(skinModelName.toLowerCase());
+      if (!path.toFile().exists()) {
+        path.toFile().mkdirs();
+      }
+      return path;
+    }
+    return null;
+  }
+
+  public static File getPresetFile(SkinModel skinModel, String fileName) {
+    Path presetModelFolder = getPresetDataFolder(skinModel);
+    if (presetModelFolder != null && fileName != null && !fileName.isEmpty()) {
+      // Remove all invalid characters from file name.
+      fileName = fileName.toLowerCase().replaceAll("[^a-z0-9/._-]", "").replace("..", "");
+      return presetModelFolder.resolve(fileName.endsWith(Constants.NPC_NBT_SUFFIX) ? fileName
+          : fileName + Constants.NPC_NBT_SUFFIX).toFile();
+    }
+    return null;
+  }
+
+  public static File getPresetFile(SkinModel skinModel, UUID uuid) {
+    return getPresetFile(skinModel, uuid.toString());
+  }
+
+  public static Stream<ResourceLocation> getPresetFilePathResourceLocations() {
+    Path presetDataFolder = getPresetDataFolder();
+    try {
+      try (Stream<Path> filesStream = Files.walk(presetDataFolder)) {
+        // Get all files with the suffix .npc.nbt and return the relative path.
+        List<ResourceLocation> filePaths =
+            filesStream.filter(path -> path.toString().endsWith(Constants.NPC_NBT_SUFFIX))
+                .filter(path -> Pattern.matches("[a-z0-9/._-]+", path.getFileName().toString()))
+                .map(path -> {
+                  ResourceLocation resourceLocation =
+                      new ResourceLocation(Constants.MOD_ID, DATA_FOLDER_NAME + '/'
+                          + presetDataFolder.relativize(path).toString().replace("\\", "/"));
+                  presetResourceLocationMap.put(resourceLocation, path);
+                  return resourceLocation;
+                }).collect(Collectors.toList());
+        return filePaths.stream();
+      }
+    } catch (IOException exception) {
+      log.error("Could not read preset data folder {}!", presetDataFolder, exception);
+    }
+
+    // Return a default or alternative stream in case of an exception
+    return Stream.empty();
+  }
+
+  public static Path getPresetsResourceLocationPath(ResourceLocation resourceLocation) {
+    return presetResourceLocationMap.get(resourceLocation);
+  }
+
+}